Move notice processing from core to demos
