+++ src/demos/index.js
@@ -2,4 +2,5 @@ export * from './file-check';
 export * from './book-package-check';
 export * from './book-packages-check';
 export * from './repo-check';
+export * from './notice-processing-functions';
 export * from './RenderProcessedResults';
